fix the potential concurrency problem and arrange code
